Revert "ARUHA-406: Avoid Kafka unavailability from cascading to Nakadi"
