+++ build.gradle
@@ -109,8 +109,6 @@ dependencies {
     compile 'org.springframework.boot:spring-boot-starter-actuator:1.3.1.RELEASE'
     compile 'org.zalando.zmon:zmon-actuator:0.9.3'
 
-    compile 'org.springframework.cloud:spring-cloud-starter-hystrix:1.1.3.RELEASE'
-
     // storage
     compile 'org.springframework.boot:spring-boot-starter-jdbc:1.3.1.RELEASE'
     compile 'org.postgresql:postgresql:9.4.1207'
