+++ src/main/java/org/zalando/nakadi/Application.java
@@ -2,10 +2,8 @@ package org.zalando.nakadi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
 
 @SpringBootApplication
-@EnableCircuitBreaker
 public class Application {
 
     public static void main(final String[] args) {
